% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adaptive_groups.R
\name{adaptive_groups}
\alias{adaptive_groups}
\alias{adaptive_groups,rda,ANY,missing-method}
\alias{adaptive_groups,rda,ANY,SpatRaster-method}
\title{Find groups of co-varying individuals/populations}
\usage{
adaptive_groups(RDA, K, env, nb_clusters)

\S4method{adaptive_groups}{rda,ANY,missing}(RDA, K, nb_clusters)

\S4method{adaptive_groups}{rda,ANY,SpatRaster}(RDA, K, env, nb_clusters)
}
\arguments{
\item{RDA}{a RDA model from which to extract loci and environmental variable scores}

\item{K}{an integer specifying the number of RDA axes to use for the clustering}

\item{env}{a stack of rasters or a data.frame with the environmental conditions of the sites to predict}

\item{nb_clusters}{an integer specifying the number of cluster to identify}
}
\value{
A \code{list} containing :
\itemize{
  \item \code{samples} : a \code{data.frame} with the cluster associated with each individual/population
  \item \code{polygons} : polygon coordinates to outline the different cluster on a 2D RDA space 
}
}
\description{
The \code{adaptive_groups} function allows the user to find groups of covarying 
individuals/populations based on their scores in the RDA space and a hkmean clustering procedure. 
\code{adaptive_groups} accommodates stack of rasters or discrete populations.
}
\details{
Identify a discrete number of adaptive groups across the 2D RDA space.
Next version will use more than two dimensions
}
\author{
Thibaut Capblancq
}
