% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CI_const.R
\name{AdjAlpha}
\alias{AdjAlpha}
\title{Adjusting alpha for one-sided CI for regression function at a point problem}
\usage{
AdjAlpha(gamma, C, X, sigma, mon_ind, lower, simlen = 10000,
  alpha = 0.05)
}
\arguments{
\item{gamma}{}

\item{C}{}

\item{X}{}

\item{sigma}{}

\item{mon_ind}{}

\item{lower}{}

\item{simlen}{}

\item{alpha}{}
}
\description{
Adjusting alpha for one-sided CI for regression function at a point problem
}
