% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CI_const.R
\name{AdjAlpha2}
\alias{AdjAlpha2}
\title{Adjusting alpha for two-sided CI for regression function at a point problem}
\usage{
AdjAlpha2(gamma, C, X, sigma, mon_ind, simlen = 10000, alpha = 0.05)
}
\arguments{
\item{gamma}{}

\item{C}{}

\item{X}{}

\item{sigma}{}

\item{mon_ind}{}

\item{simlen}{}

\item{alpha}{}
}
\description{
Adjusting alpha for two-sided CI for regression function at a point problem
}
