% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CI_const.R
\name{AdjAlpha_RD}
\alias{AdjAlpha_RD}
\title{(One-sided) Optimal "tau"}
\usage{
AdjAlpha_RD(gamma, C, Xt, Xc, mon_ind, sigma_t, sigma_c, lower,
  simlen = 10000, alpha = 0.05)
}
\arguments{
\item{gamma}{length two vector of gammas (\eqn{(\gamma_1, \gamma_2)'})}

\item{C}{length two vector of Cs (\eqn{(C_1, C_2)'})}

\item{Xt}{\eqn{n_t \times k} design matrix for the treated units}

\item{Xc}{\eqn{n_c \times k} design matrix for the control units}

\item{mon_ind}{indice of the monotone variables}

\item{sigma_t}{standard deviation of the error term for the treated units
(either length 1 or \eqn{n_t})}

\item{sigma_c}{standard deviation of the error term for the control units}

\item{lower}{}

\item{simlen}{}

\item{alpha}{}

\item{Yt}{length \eqn{n_t} of outcome variables for the treated units}

\item{Yc}{length \eqn{n_c} of outcome variables for the control units}
}
\description{
Calculates the optimal "tau"
}
