% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCres_comb.R
\name{adpt.res.gen}
\alias{adpt.res.gen}
\title{Length Generation for One-sided CIs}
\usage{
adpt.res.gen(
  eps,
  CI.method,
  th,
  spec.etc,
  which.fun,
  n,
  sig.scale,
  which.spec = 1
)
}
\arguments{
\item{eps}{N(0,1) random noises.}

\item{CI.method}{method to be used to generate the CI.}

\item{th}{true RD parameter value.}

\item{spec.etc}{list of other specification parameters.}

\item{which.fun}{which regression function to evaluate; supports \code{1:6}.}

\item{n}{sample size.}

\item{sig.scale}{scaling parameter for the standard deviation.}

\item{which.spec}{which design is used; the default is \code{which.spec = 1}.}
}
\value{
vector of excess lengths and coverage indicators corresponding to
true Lipschitz coefficient values contained in \code{spec.etc$C.true.vec}.
}
\description{
Calculate excess lengths of one-sided CIs for different values of true Lipschitz coefficients.
}
\details{
Modification for multi-dim hasn't been made yet.
}
\examples{
spec.etc <- list(alpha = 0.05, mon_ind = 1,
C.small = 1, C.large = 3, se.method.RDH = "nn", se.initial.RDH = "nn", M.RDH = 1,
se.method = "nn", se.init = "Silverman", t.dir = "left", C.l = 1/2, C.u = 2, C = 3,
C.true.vec = seq(1/2, 3, length.out = 2), adpt.orc = TRUE)
n <- 100
th <- 1
eps <- rnorm(n)
which.fun <- 1
sig.scale <- 1/2
CI.method <- "adpt.one"
adpt.res.gen(eps, CI.method, th, spec.etc, which.fun, n, sig.scale)
spec.etc$adpt.orc <- FALSE
adpt.res.gen(eps, CI.method, th, spec.etc, which.fun, n, sig.scale)
}
