% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_calc.R
\name{bmodsol}
\alias{bmodsol}
\title{Between modulus}
\usage{
bmodsol(delta, gamma, C, X, mon_ind, sigma = 1)
}
\arguments{
\item{delta}{}

\item{gamma}{}

\item{C}{}

\item{X}{}

\item{mon_ind}{}
}
\description{
Calculates the between modulus for the regression function at a point
problem
}
