% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CI_const_RD.R
\name{bw_mod}
\alias{bw_mod}
\title{Between Modulus Calculation}
\usage{
bw_mod(delta, Cj, Cbar, Xt, Xc, mon_ind, sigma_t, sigma_c)
}
\arguments{
\item{delta}{a nonnegative value of \eqn{\delta}.}

\item{Cj}{the smoothness parameter aiming to adapt to.}

\item{Cbar}{the largest smoothness parameter.}

\item{Xt}{\eqn{n_t} by \eqn{k} design matrix for the treated units.}

\item{Xc}{\eqn{n_c} by \eqn{k} design matrix for the control units.}

\item{mon_ind}{index number for monotone variables.}

\item{sigma_t}{standard deviation of the error term for the treated units
(either length 1 or \eqn{n_t}).}

\item{sigma_c}{standard deviation of the error term for the control units
(either length 1 or \eqn{n_c}).}
}
\value{
a list of two values, \code{h_jt} and \code{h_jc}.
}
\description{
Calculates the between moduli \eqn{b_{jt}} and \eqn{b_{jc}}
for a given value of \eqn{\delta} and \eqn{C_j}.
}
\examples{
n <- 500
d <- 2
X <- matrix(rnorm(n * d), nrow = n, ncol = d)
tind <- X[, 1] < 0 & X[, 2] < 0
Xt <- X[tind == 1, ,drop = FALSE]
Xc <- X[tind == 0, ,drop = FALSE]
mon_ind <- c(1, 2)
sigma <- rnorm(n)^2 + 1
sigma_t <- sigma[tind == 1]
sigma_c <- sigma[tind == 0]
bw_mod(1, 1/2, 1, Xt, Xc, mon_ind, sigma_t, sigma_c)
bw_mod(1, 1/2, Inf, Xt, Xc, mon_ind, sigma_t, sigma_c)
}
