% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CI_const_RD.R
\name{c_hat_lower_RD}
\alias{c_hat_lower_RD}
\title{Lower Adaptive CI for the RD Parameter}
\usage{
c_hat_lower_RD(
  delta,
  Cj,
  Cbar,
  Xt,
  Xc,
  mon_ind,
  sigma_t,
  sigma_c,
  Yt,
  Yc,
  tau,
  ht,
  hc
)
}
\arguments{
\item{delta}{a nonegative scalar value:
it can be left unspecified if \code{ht} and \code{hc} are specified.}

\item{Cj}{the smoothness parameter aiming to adapt to.}

\item{Cbar}{the largest smoothness parameter.}

\item{Xt}{\eqn{n_t} by \eqn{k} design matrix for the treated units.}

\item{Xc}{\eqn{n_c} by \eqn{k} design matrix for the control units.}

\item{mon_ind}{index number for monotone variables.}

\item{sigma_t}{standard deviation of the error term for the treated units
(either length 1 or \eqn{n_t}).}

\item{sigma_c}{standard deviation of the error term for the control units
(either length 1 or \eqn{n_c}).}

\item{Yt}{outcome value for the treated group observations.}

\item{Yc}{outcome value for the control group observations.}

\item{tau}{desired level of non-coverage probability.}

\item{ht}{the modulus value for the treated observations;
it can be left unspecified if \code{delta} is specified.}

\item{hc}{the modulus value for the control observations;
it can be left unspecified if \code{delta} is specified.}
}
\value{
the value of the lower end point of the adaptive CI.
}
\description{
Calculates the lower end point of the adaptive CI for the RD parameter.
}
\examples{
n <- 500
d <- 2
X <- matrix(rnorm(n * d), nrow = n, ncol = d)
tind <- X[, 1] < 0 & X[, 2] < 0
Xt <- X[tind == 1, ,drop = FALSE]
Xc <- X[tind == 0, ,drop = FALSE]
mon_ind <- c(1, 2)
sigma <- rnorm(n)^2 + 1
sigma_t <- sigma[tind == 1]
sigma_c <- sigma[tind == 0]
Yt = 1 + rnorm(length(sigma_t), mean = 0, sd = sigma_t)
Yc = rnorm(length(sigma_c), mean = 0, sd = sigma_c)
c_hat_lower_RD(1, 1/2, 1, Xt, Xc, mon_ind, sigma_t, sigma_c,
Yt, Yc, 0.05)
c_hat_lower_RD(1, 1/2, Inf, Xt, Xc, mon_ind, sigma_t, sigma_c,
Yt, Yc, 0.05)
}
