% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CI_intsec_RD.R
\name{CI_adpt_L}
\alias{CI_adpt_L}
\title{Lower Adaptive Confidence Interval}
\usage{
CI_adpt_L(
  Cvec,
  Xt,
  Xc,
  mon_ind,
  sigma_t,
  sigma_c,
  Yt,
  Yc,
  alpha,
  delta_init = 1.96,
  hmat_init,
  hmat
)
}
\arguments{
\item{Cvec}{a sequence of smoothness parameters}

\item{Xt}{\eqn{n_t} by \eqn{k} design matrix for the treated units.}

\item{Xc}{\eqn{n_c} by \eqn{k} design matrix for the control units.}

\item{mon_ind}{index number for monotone variables.}

\item{sigma_t}{standard deviation of the error term for the treated units
(either length 1 or \eqn{n_t}).}

\item{sigma_c}{standard deviation of the error term for the control units
(either length 1 or \eqn{n_c}).}

\item{Yt}{outcome value for the treated group observations.}

\item{Yc}{outcome value for the control group observations.}

\item{alpha}{desired upper quantile value.}

\item{delta_init}{the value of \eqn{\delta} to be used in simulating the quantile;
theoretically, its value does not matter asymptotically. Its default value is 1.96.}

\item{hmat_init}{the matrix of modulus values corresponding to \code{delta_init}
and \code{Cvec}; it can be left unspecified.}

\item{hmat}{the matrix of modulus values corresponding to
the optimal \eqn{\delta} and \code{Cvec}; it can be left unspecified.}
}
\value{
a list components \code{CI} which gives the value of
the lower end of the confidence interval, \code{CI_vec} which gives the values of
all CIs used to compute the adaptive CI, and \code{tau_sol} which gives the value of
non-coverage probability used to compute individual CIs.
}
\description{
Calculates the value of the lower end of the adaptive confidence interval.
}
\examples{
n <- 500
d <- 2
X <- matrix(rnorm(n * d), nrow = n, ncol = d)
tind <- X[, 1] > 0 & X[, 2] > 0
Xt <- X[tind == 1, ,drop = FALSE]
Xc <- X[tind == 0, ,drop = FALSE]
mon_ind <- c(1, 2)
sigma <- rnorm(n)^2 + 1
sigma_t <- sigma[tind == 1]
sigma_c <- sigma[tind == 0]
Yt = 1 + rnorm(length(sigma_t), mean = 0, sd = sigma_t)
Yc = rnorm(length(sigma_c), mean = 0, sd = sigma_c)
CI_adpt_L((1:5)/5, Xt, Xc, mon_ind, sigma_t, sigma_c, Yt, Yc, 0.05)
}
