% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CI_const.R
\name{CI_gen}
\alias{CI_gen}
\title{Title}
\usage{
CI_gen(Yt, Yc, Xt, Xc, C, C_max = max(C), mon_ind, sigma_t, sigma_c,
  lower, two_sided = F, alpha = 0.05, modres = NULL)
}
\arguments{
\item{Yt}{}

\item{Yc}{}

\item{Xt}{}

\item{Xc}{}

\item{C}{}

\item{C_max}{}

\item{mon_ind}{}

\item{sigma_t}{}

\item{sigma_c}{}

\item{lower}{}

\item{two_sided}{}

\item{alpha}{}

\item{modres}{}
}
\description{
Title
}
