% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sim_fcns.R
\name{CI_gen_met}
\alias{CI_gen_met}
\title{CI Generation Corresponding to the Specified Method}
\usage{
CI_gen_met(
  met = c("Ex", "Csvtv", "rdr"),
  Xt,
  Xc,
  mon_ind,
  sigma_t,
  sigma_c,
  Yt,
  Yc,
  alpha,
  trueC = NULL,
  C_len = NULL,
  rdr_swap = TRUE,
  rdr_met = 3
)
}
\arguments{
\item{met}{CI generation method to use, among \code{c("Ex", "Csvtv", "rdr")} so far.}

\item{Xt}{\eqn{n_t} by \eqn{k} design matrix for the treated units.}

\item{Xc}{\eqn{n_c} by \eqn{k} design matrix for the control units.}

\item{mon_ind}{index number for monotone variables.}

\item{sigma_t}{standard deviation of the error term for the treated units
(either length 1 or \eqn{n_t}).}

\item{sigma_c}{standard deviation of the error term for the control units
(either length 1 or \eqn{n_c}).}

\item{Yt}{outcome value for the treated group observations.}

\item{Yc}{outcome value for the control group observations.}

\item{alpha}{a desired level of non-coverage.}

\item{trueC}{a vector of true smoothness parameters used in simulations.}

\item{C_len}{the number of smoothness parameters to adapt to.}

\item{rdr_swap}{\code{TRUE} if swapped value of \code{rdrobust} result will be used;
set \code{TRUE} if \code{x < 0} is treated.}

\item{rdr_met}{which row of the \code{rdrobust} result matrix will be used; the default is 3.}
}
\value{
a vector of lower and upper ends of the CI.
}
\description{
CI Generation Corresponding to the Specified Method
}
\examples{
obs_data <- rdadapt::gen_obs(100, 1, 1/2, "Lin", "het", "equi")
Xt <- obs_data$Xt
Xc <- obs_data$Xc
sigma_t <- obs_data$sigma_t
sigma_c <- obs_data$sigma_c
Yt <- obs_data$Yt
Yc <- obs_data$Yc
trueC <- (1:5)/5
C_len <- 5
CI_gen_met("Ex", Xt, Xc, c(1), sigma_t, sigma_c, Yt, Yc, 0.05, trueC, C_len)
CI_gen_met("rdr", Xt, Xc, c(1), sigma_t, sigma_c, Yt, Yc, 0.05)
}
