% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CI_const.R
\name{CI_length}
\alias{CI_length}
\title{Minimax CI length for regression function at a point}
\usage{
CI_length(b, gamma, C, X, mon_ind, sigma = 1, alpha = 0.05)
}
\arguments{
\item{b}{}

\item{gamma}{}

\item{C}{}

\item{X}{}

\item{mon_ind}{}

\item{sigma}{}
}
\description{
This function calculates the minimax CI length for the regression function
at a point problem
}
