% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CI_mm.R
\name{CI_length_RD}
\alias{CI_length_RD}
\title{Minimum CI Half-length}
\usage{
CI_length_RD(b, C, Xt, Xc, mon_ind, sigma_t, sigma_c, alpha)
}
\arguments{
\item{b}{point where the inverse modulus is evaluated at.}

\item{C}{a scalar smoothness parameter.}

\item{Xt}{\eqn{n_t} by \eqn{d} design matrix for the treated units.}

\item{Xc}{\eqn{n_c} by \eqn{d} design matrix for the control units.}

\item{mon_ind}{index number for monotone variables.}

\item{sigma_t}{standard deviation of the error term for the treated units
(either length 1 or \eqn{n_t}).}

\item{sigma_c}{standard deviation of the error term for the control units
(either length 1 or \eqn{n_c}).}

\item{alpha}{the desired level of non-coverage probability.}
}
\value{
the length of CI corresponding to the modulus value \code{b}
}
\description{
Minimum CI Half-length
}
\examples{
n <- 500
d <- 2
X <- matrix(rnorm(n * d), nrow = n, ncol = d)
tind <- X[, 1] > 0 & X[, 2] > 0
Xt <- X[tind == 1, ,drop = FALSE]
Xc <- X[tind == 0, ,drop = FALSE]
C <- 1/2
mon_ind <- c(1, 2)
sigma <- rnorm(n)^2 + 1
sigma_t <- sigma[tind == 1]
sigma_c <- sigma[tind == 0]
CI_length_RD(1, C, Xt, Xc, mon_ind, sigma_t , sigma_c, 0.05)
}
