% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CI_const.R
\name{CI_one_sd}
\alias{CI_one_sd}
\title{One-sided adaptive CI}
\usage{
CI_one_sd(bpairmat, delta, gamma, C, X, mon_ind, sigma, Y, lower, al,
  dp = FALSE, maxQ = FALSE, simlen = 50000, qres = FALSE,
  alpha = 0.05)
}
\arguments{
\item{bpairmat}{}

\item{delta}{}

\item{gamma}{}

\item{C}{}

\item{X}{}

\item{mon_ind}{}

\item{sigma}{}

\item{Y}{}

\item{lower}{}

\item{al}{}

\item{dp}{indicator of whether to direct power}

\item{maxQ}{}

\item{simlen}{}

\item{qres}{}

\item{alpha}{}
}
\description{
This function calculates the one cided adaptive CI for the regression
function at a point problem as given in Kwon and Kwon (2019)
}
