% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CI_const.R
\name{CI_one_sd_RD}
\alias{CI_one_sd_RD}
\title{(One-sided) Adaptive CI for RDD}
\usage{
CI_one_sd_RD(bpairmat, dmat, gamma, C, Xt, Xc, mon_ind, sigma_t, sigma_c,
  yt, yc, lower, al, Dir = FALSE)
}
\arguments{
\item{gamma}{length two vector of gammas (\eqn{(\gamma_1, \gamma_2)'})}

\item{C}{length two vector of Cs (\eqn{(C_1, C_2)'})}

\item{Xt}{\eqn{n_t \times k} design matrix for the treated units}

\item{Xc}{\eqn{n_c \times k} design matrix for the control units}

\item{mon_ind}{indice of the monotone variables}

\item{sigma_t}{standard deviation of the error term for the treated units
(either length 1 or \eqn{n_t})}

\item{sigma_c}{standard deviation of the error term for the control units}

\item{Yt}{length \eqn{n_t} of outcome variables for the treated units}

\item{Yc}{length \eqn{n_c} of outcome variables for the control units}
}
\description{
Calculates the (one-sided) adaptive CI for RDD
}
