% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CI_intsec_RD.R
\name{cov_calc}
\alias{cov_calc}
\title{Covariance Between Two Estimators}
\usage{
cov_calc(
  delta,
  Cj,
  Ck,
  Cbar,
  Xt,
  Xc,
  mon_ind,
  sigma_t,
  sigma_c,
  ht_j,
  hc_j,
  ht_k,
  hc_k
)
}
\arguments{
\item{delta}{a nonegative scalar value;
it can be left unspecified if
(\code{ht_j}, \code{ht_k}, \code{hc_j}, \code{hc_k}) are specified.}

\item{Cj}{the smoothness parameter corresponding to the first estimator.}

\item{Ck}{the smoothness parameter corresponding to the second estimator.}

\item{Cbar}{the largest smoothness parameter.}

\item{Xt}{\eqn{n_t} by \eqn{k} design matrix for the treated units.}

\item{Xc}{\eqn{n_c} by \eqn{k} design matrix for the control units.}

\item{mon_ind}{index number for monotone variables.}

\item{sigma_t}{standard deviation of the error term for the treated units
(either length 1 or \eqn{n_t}).}

\item{sigma_c}{standard deviation of the error term for the control units
(either length 1 or \eqn{n_c}).}

\item{ht_j}{the modulus value for the treated observations,
corresponding to the first smoothness parameter;
it can be left unspecified if \code{delta} is specified.}

\item{hc_j}{the modulus value for the contorl observations,
corresponding to the first smoothness parameter;
it can be left unspecified if \code{delta} is specified.}

\item{ht_k}{the modulus value for the treated observations,
corresponding to the second smoothness parameter;
it can be left unspecified if \code{delta} is specified.}

\item{hc_k}{the modulus value for the control observations,
corresponding to the second smoothness parameter;
it can be left unspecified if \code{delta} is specified.}
}
\value{
a scalar covariance value.
}
\description{
Calculates the covariance between two estimators,
\eqn{Lhat_j(\delta)} and \eqn{Lhat_k(\delta)}.
}
\examples{
n <- 500
d <- 2
X <- matrix(rnorm(n * d), nrow = n, ncol = d)
tind <- X[, 1] < 0 & X[, 2] < 0
Xt <- X[tind == 1, ,drop = FALSE]
Xc <- X[tind == 0, ,drop = FALSE]
mon_ind <- c(1, 2)
sigma <- rnorm(n)^2 + 1
sigma_t <- sigma[tind == 1]
sigma_c <- sigma[tind == 0]
cov_calc(1, 0.2, 0.4, 1, Xt, Xc, mon_ind, sigma_t, sigma_c)
cov_calc(1, 0.2, 0.4, Inf, Xt, Xc, mon_ind, sigma_t, sigma_c)
}
