% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CI_intsec_RD.R
\name{cov_mat_calc}
\alias{cov_mat_calc}
\title{Covariance Matrix Calculation for Multiple Estimators}
\usage{
cov_mat_calc(delta, Cvec, Cbar, Xt, Xc, mon_ind, sigma_t, sigma_c, hmat)
}
\arguments{
\item{delta}{a nonegative scalar value;
it can be left unspecified if
\code{hmat} is specified.}

\item{Cvec}{a sequence of smoothness parameters}

\item{Cbar}{the Lipschitz coefficient for the largest function space we consider}

\item{Xt}{\eqn{n_t} by \eqn{k} design matrix for the treated units.}

\item{Xc}{\eqn{n_c} by \eqn{k} design matrix for the control units.}

\item{mon_ind}{index number for monotone variables.}

\item{sigma_t}{standard deviation of the error term for the treated units
(either length 1 or \eqn{n_t}).}

\item{sigma_c}{standard deviation of the error term for the control units
(either length 1 or \eqn{n_c}).}

\item{hmat}{a \eqn{J} by {2} matrix of modulus values;
it can be left unspecified if \code{delta} is specified.}
}
\value{
a \eqn{J} by \eqn{J} covariance matrix.
}
\description{
Calculates the covariance matrix of \eqn{J} estimators,
\eqn{Lhat_1(\delta)},..., \eqn{Lhat_J(\delta)}.
}
\examples{
n <- 500
d <- 2
X <- matrix(rnorm(n * d), nrow = n, ncol = d)
tind <- X[, 1] < 0 & X[, 2] < 0
Xt <- X[tind == 1, ,drop = FALSE]
Xc <- X[tind == 0, ,drop = FALSE]
mon_ind <- c(1, 2)
sigma <- rnorm(n)^2 + 1
sigma_t <- sigma[tind == 1]
sigma_c <- sigma[tind == 0]
cov_mat_calc(1, (1:5)/5, 1, Xt, Xc, mon_ind, sigma_t, sigma_c)
cov_mat_calc(1, (1:5)/5, Inf, Xt, Xc, mon_ind, sigma_t, sigma_c)
}
