% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CI_mm.R
\name{cv_a}
\alias{cv_a}
\title{Quantile of Folded Normal Distribution}
\usage{
cv_a(t, alpha)
}
\arguments{
\item{t}{the non-centrality parameter.}

\item{alpha}{probability}
}
\value{
the quantile value
}
\description{
Calculates the \eqn{1 - \alpha} quantile of \eqn{|N(t, 1)|}.
}
\examples{
cv_a(1, 0.05)
}
