% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cvec_choice.R
\name{dist.J}
\alias{dist.J}
\title{Distance Measure Function (of J)}
\usage{
dist.J(
  C_l,
  C_u,
  C,
  Xt,
  Xc,
  mon_ind,
  sigma_t,
  sigma_c,
  alpha,
  J,
  n_grid = 10,
  ratio = TRUE,
  p = Inf,
  n_sim = 10^5
)
}
\arguments{
\item{C_l}{lower end of the adaptation range.}

\item{C_u}{upper end of the adaptation range.}

\item{C}{the Lipschitz coefficient for the function space we consider.}

\item{Xt}{\eqn{n_t} by \eqn{k} design matrix for the treated units.}

\item{Xc}{\eqn{n_c} by \eqn{k} design matrix for the control units.}

\item{mon_ind}{index number for monotone variables.}

\item{sigma_t}{standard deviation of the error term for the treated units
(either length 1 or \eqn{n_t}).}

\item{sigma_c}{standard deviation of the error term for the control units
(either length 1 or \eqn{n_c}).}

\item{alpha}{desired upper quantile value.}

\item{J}{the value of J}

\item{n_grid}{number of grid points to evaluate the lengths.}

\item{ratio}{the ratio measure is used if \code{TRUE};
otherwise, the difference measure is used.}

\item{p}{the order of \eqn{l_1}-norm; the default is \code{Inf}.}

\item{n_sim}{number of simulated observations to calculate the
expectation of the minimum of multivariate normal random variables;
the default is \code{n_sim = 10^5}.}
}
\value{
the distance measure
}
\description{
Distance Measure Function (of J)
}
\examples{
n <- 500
d <- 2
X <- matrix(rnorm(n * d), nrow = n, ncol = d)
tind <- X[, 1] < 0 & X[, 2] < 0
Xt <- X[tind == 1, ,drop = FALSE]
Xc <- X[tind == 0, ,drop = FALSE]
mon_ind <- c(1, 2)
sigma <- rnorm(n)^2 + 1
sigma_t <- sigma[tind == 1]
sigma_c <- sigma[tind == 0]
dist.J(0.1, 1, 2, Xt, Xc, mon_ind, sigma_t, sigma_c, 0.05, 2)
}
