% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cvec_choice.R
\name{dist_l}
\alias{dist_l}
\title{Distance Between Two Length Functions}
\usage{
dist_l(l_adpt_vec, l_orc_vec, ratio = TRUE, p = Inf)
}
\arguments{
\item{l_adpt_vec}{vector of worst case lengths for the adaptive procedure.}

\item{l_orc_vec}{vector of worst case lengths for the oracle procedure.}

\item{ratio}{the ratio measure is used if \code{TRUE};
otherwise, the difference measure is used.}

\item{p}{the order of \eqn{l_1}-norm; the default is \code{Inf}.}
}
\description{
Calculates \eqn{\Delta(l_1, l_2)} in the paper (Ver20201104).
}
\examples{
l_adpt_vec <- 1:10
l_orc_vec <- (1:10) - 1/2
dist_l(l_adpt_vec, l_orc_vec)
}
