% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Reg_fcns.R
\name{f.1}
\alias{f.1}
\title{Function \eqn{f_1}}
\usage{
f.1(x, C, th)
}
\arguments{
\item{x}{vector of regressor values.}

\item{C}{smoothness parameter.}

\item{th}{true RD parameter.}
}
\value{
vector of regression function values
}
\description{
Function \eqn{f_1} in the version 20201104
}
\examples{
x <- runif(100, -1, 1)
f.1(x, 1, 1)
}
