% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Reg_fcns.R
\name{f.2}
\alias{f.2}
\title{Function \eqn{f_2}}
\usage{
f.2(x, C, th, b1 = 1/3, b2 = 2/3)
}
\arguments{
\item{x}{vector of regressor values.}

\item{C}{smoothness parameter.}

\item{th}{true RD parameter.}

\item{b1}{the first knot value; the default is \code{b1 = 1/3}.}

\item{b2}{the second knot value; the default is \code{b2 = 2/3}.}
}
\value{
vector of regression function values
}
\description{
Function \eqn{f_2} in the version 20201104
}
\examples{
x <- runif(100, -1, 1)
f.2(x, 1, 1)
}
