% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Reg_fcns.R
\name{f.4}
\alias{f.4}
\title{Function \eqn{f_4}}
\usage{
f.4(x, C, th)
}
\arguments{
\item{x}{vector of regressor values.}

\item{C}{smoothness parameter.}

\item{th}{true RD parameter.}
}
\value{
vector of regression function values
}
\description{
Function \eqn{f_4} in the version 20201105
}
\examples{
x <- runif(100, -1, 1)
f.4(x, 1, 1)
}
