% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Reg_fcns.R
\name{f.5}
\alias{f.5}
\title{Function \eqn{f_5}}
\usage{
f.5(x.mat, C, th)
}
\arguments{
\item{x.mat}{n by 2 matrix}

\item{C}{smoothness parameter.}

\item{th}{true RD parameter.}
}
\value{
vector of regression function values
}
\description{
Function \eqn{f_5} in the version 20201104
}
\examples{
x.mat <- cbind(runif(100, -1, 1), runif(100, -1, 1))
f.5(x.mat, 1, 1)
}
