% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_calc.R
\name{f2subf1_fun}
\alias{f2subf1_fun}
\title{Calculate \eqn{f_2^* - f_1^*}}
\usage{
f2subf1_fun(b, gamma, C, X, mon_ind, swap = FALSE)
}
\arguments{
\item{b}{point where the inverse modulus is evaluated at}

\item{gamma}{length two vector of gammas (\eqn{(\gamma_1, \gamma_2)'})}

\item{C}{length two vector of Cs (\eqn{(C_1, C_2)'})}

\item{X}{n\eqn{\times}k design matrix, the kernel is evaluated at each X[i, ]}

\item{mon_ind}{indice of the monotone variables}

\item{swap}{indicator of whether to swap the \eqn{\gamma} and \eqn{C}.}
}
\description{
This function calculates \eqn{f_2^* - f_1^*} that correponds to
\eqn{\omega^{-1}(b, \Lambda_{\mathcal{V}+}(\gamma_1, C_1),\Lambda_{\mathcal{V}+}(\gamma_2, C_2)) }
}
