% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomic_offset.R
\name{genomic_offset}
\alias{genomic_offset}
\alias{genomic_offset,rda,ANY,SpatRaster,ANY,ANY,missing-method}
\alias{genomic_offset,rda,ANY,data.frame,ANY,missing,missing-method}
\alias{genomic_offset,rda,ANY,data.frame,missing,missing,ANY-method}
\title{Predict temporal or geographic genomic offset}
\usage{
genomic_offset(
  RDA,
  K,
  env_pres,
  env_fut,
  env_mask = NULL,
  env_gar,
  method = "loadings"
)

\S4method{genomic_offset}{rda,ANY,SpatRaster,ANY,ANY,missing}(RDA, K, env_pres, env_fut, env_mask = NULL, method = "loadings")

\S4method{genomic_offset}{rda,ANY,data.frame,ANY,missing,missing}(RDA, K, env_pres, env_fut, method = "loadings")

\S4method{genomic_offset}{rda,ANY,data.frame,missing,missing,ANY}(RDA, K, env_pres, env_gar, method = "loadings")
}
\arguments{
\item{RDA}{a \code{RDA} model from which to extract loci and environmental variable scores}

\item{K}{an \code{integer} specifying the number of RDA axes to use for the projection}

\item{env_pres}{a \code{RasterStack} object or a \code{data.frame} with the environmental 
conditions in the present}

\item{env_fut}{a \code{RasterStack} object or a \code{data.frame} with the environmental 
conditions in the future}

\item{env_mask}{(\emph{optional, default} \code{NULL}) \cr a \code{Raster} object to limit 
the projection to a specific area}

\item{method}{(\emph{default} \code{'loadings'}) \cr a \code{character} defining whether 
the function is to use weighted averages (scaling type 1, \code{loadings}) or linear 
combinations (scaling type 2, \code{predict}) of the projected environmental variables 
to predict site scores (i.e., adaptive index)}
}
\value{
A \code{list} containing :
\itemize{
  \item \code{genomic_offset} : a \code{RasterStack} or a \code{data.frame} containing 
  the genomic offset predictions for the \code{K} first RDA axes, as well as the overall 
  genomic offset prediction
  \item \code{weights} : the weights associated with each RDA axis used for the predictions
}
}
\description{
The \code{genomic_offset} function allows the user to predict genomic offset 
from a RDA model. The \code{genomic_offset} function can estimate both temporal or spatial 
genomic offset and accommodates raster data or discrete populations.
}
\details{
This RDA-based method to predict \emph{genomic offset} is relatively simple. 
RDA is first used to predict the \emph{optimal adaptive genetic composition} for each 
environmental pixel under consideration (see \code{\link{adaptive_index}} function), 
using both current and future environmental conditions. 
The \emph{euclidean distance} between these two predictions in the RDA space provides an 
estimate of the change in genetic composition that would be required to track climate change.
}
\seealso{
\code{\link{adaptive_index}}
}
\author{
Thibaut Capblancq
}
\keyword{RDA,}
\keyword{adaptive}
\keyword{genomic}
\keyword{index,}
\keyword{offset,}
\keyword{projection}
