% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_calc_RD.R
\name{invmod}
\alias{invmod}
\title{Inverse Modulus}
\usage{
invmod(b, C_pair, X, mon_ind, sigma, swap = FALSE)
}
\arguments{
\item{b}{point where the inverse modulus is evaluated at.}

\item{C_pair}{a pair of smoothness parameters \eqn{(C, C')}.}

\item{X}{n by d design matrix.}

\item{mon_ind}{index of the monotone variables.}

\item{sigma}{standard deviation of the error term (either length 1 or n).}

\item{swap}{indicator for whether we take (C', C) instead of (C, C').}
}
\value{
the value of inverse modulus given \eqn{\omega = b}.
}
\description{
Calculates the inverse modulus for the regression function at a point
problem. More specifically, this calcultes
\eqn{\omega^{-1}(b, \Lambda_{V+}(C),\Lambda_{V+}(C')) }
}
\examples{
X <- matrix(c(1, -2, -3, 4, 5, -6), nrow = 3, ncol = 2)
mon_ind <- c(1, 2)
C_pair <- c(0.5, 1)
sigma <- c(1, 2, 1)
invmod(1, C_pair, X, mon_ind, sigma)
}
