% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_fcn_RD.R
\name{K_fun}
\alias{K_fun}
\title{Kernel Function for Adaptive Procedures}
\usage{
K_fun(b, C_pair, X, mon_ind, swap = FALSE)
}
\arguments{
\item{b}{a scalar "modulus of continuity" parameter;
can take the value of 0.}

\item{C_pair}{a pair of smoothness parameters \eqn{(C, C')};
can take the values of \code{Inf}.}

\item{X}{data matrix.}

\item{mon_ind}{index number for monotone variables, with length
\code{ncol(X)}.}

\item{swap}{we take (C', C) instead of (C, C') if \code{swap = TRUE}.}
}
\value{
A numeric vector with the length equal to \code{nrow(X)};
if b = 0, returns 0 vector.
}
\description{
Calculates \eqn{K(x/b; C, C')} in our previous draft.
}
\examples{
X <- matrix(c(1, -2, -3, 4, 5, -6), nrow = 3, ncol = 2)
mon_ind <- c(1, 2)
C_pair <- c(0.5, 1)
K_fun(1, C_pair, X, mon_ind)
K_fun(1, c(Inf, 1), X, mon_ind)
}
