% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cvec_choice.R
\name{l_orc}
\alias{l_orc}
\title{Oracle Excess Length}
\usage{
l_orc(Cpr, C, Xt, Xc, mon_ind, sigma_t, sigma_c, alpha)
}
\arguments{
\item{Cpr}{\eqn{C'} in the paper}

\item{C}{the Lipschitz coefficient for the function space we consider}

\item{Xt}{\eqn{n_t} by \eqn{k} design matrix for the treated units.}

\item{Xc}{\eqn{n_c} by \eqn{k} design matrix for the control units.}

\item{mon_ind}{index number for monotone variables.}

\item{sigma_t}{standard deviation of the error term for the treated units
(either length 1 or \eqn{n_t}).}

\item{sigma_c}{standard deviation of the error term for the control units
(either length 1 or \eqn{n_c}).}

\item{alpha}{desired upper quantile value.}
}
\description{
Oracle Excess Length
}
\examples{
n <- 500
d <- 2
X <- matrix(rnorm(n * d), nrow = n, ncol = d)
tind <- X[, 1] < 0 & X[, 2] < 0
Xt <- X[tind == 1, ,drop = FALSE]
Xc <- X[tind == 0, ,drop = FALSE]
mon_ind <- c(1, 2)
sigma <- rnorm(n)^2 + 1
sigma_t <- sigma[tind == 1]
sigma_c <- sigma[tind == 0]
l_orc(1/4, 2, Xt, Xc, mon_ind, sigma_t, sigma_c, 0.05)
}
