% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CI_const.R
\name{Lhat_fun}
\alias{Lhat_fun}
\title{Estimator used in constructing the minimax (or adaptive) CIs}
\usage{
Lhat_fun(b, gamma, C, X, mon_ind, sigma, Y, swap = FALSE)
}
\arguments{
\item{b}{}

\item{gamma}{}

\item{C}{}

\item{X}{}

\item{mon_ind}{}

\item{sigma}{}

\item{Y}{}

\item{swap}{}
}
\description{
Calculates the estimator used in constructing then minimax (or adaptive) CIs
for the regression function at a point problem. Again, the
estimator corresponds to \eqn{\omega^{-1}(b,
\Lambda_{\mathcal{V}+}(\gamma_2, C_2),\Lambda_{\mathcal{V}+}(\gamma_1, C_1))
}
}
