% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CI_const.R
\name{Lhat_RD_fun}
\alias{Lhat_RD_fun}
\title{Estimator used in constructing the minimax CI for RDD}
\usage{
Lhat_RD_fun(b = NULL, bt = NULL, bc = NULL, gamma, C, Xt, Xc, Yt, Yc,
  mon_ind, sigma_t, sigma_c, swap = FALSE)
}
\arguments{
\item{b}{point where the inverse modulus is evaluated at}

\item{gamma}{length two vector of gammas (\eqn{(\gamma_1, \gamma_2)'})}

\item{C}{length two vector of Cs (\eqn{(C_1, C_2)'})}

\item{Xt}{\eqn{n_t \times k} design matrix for the treated units}

\item{Xc}{\eqn{n_c \times k} design matrix for the control units}

\item{Yt}{length \eqn{n_t} of outcome variables for the treated units}

\item{Yc}{length \eqn{n_c} of outcome variables for the control units}

\item{mon_ind}{indice of the monotone variables}

\item{sigma_t}{standard deviation of the error term for the treated units
(either length 1 or \eqn{n_t})}

\item{sigma_c}{standard deviation of the error term for the control units}

\item{swap}{indiactor of whether to swap the parameter spaces}
}
\description{
Calculates the estimator that is used to construct the adaptive (or minimax)
CI for RDD. Again, the estimator correponds to the inverse modulus
\eqn{\omega^{-1}(b, \Lambda_{\mathcal{V}+}(\gamma_1,
C_1),\Lambda_{\mathcal{V}+}(\gamma_2, C_2)). }
}
