% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CI_const.R
\name{Lhatfun}
\alias{Lhatfun}
\title{Estimator used in constructing adaptive CIs for RDD}
\usage{
Lhatfun(b, C1, C2, g1, g2, X, mon_ind, sigma, y)
}
\arguments{
\item{b}{point where the inverse modulus is evaluated at}

\item{C1}{\eqn{C_1}}

\item{C2}{\eqn{C_2}}

\item{g1}{\eqn{\gamma_1}}

\item{g2}{\eqn{\gamma_2}}

\item{X}{design matrix}
}
\description{
Calculates the estimator that is used to construct the adaptive
CI for RDD. The estimator corresponds to the inverse modulus
\eqn{\omega^{-1}(b, \Lambda_{\mathcal{V}+}(\gamma_2, C_2),\Lambda_{\mathcal{V}+}(\gamma_1, C_1)) }
}
