% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loci_module.R
\name{loci_module}
\alias{loci_module}
\title{Loci clustering}
\usage{
loci_module(RDA, K)
}
\arguments{
\item{RDA}{blabla}

\item{K}{blabla}
}
\value{
A \code{data.frame} containing :
\itemize{
  \item \code{p.values} : blabla
  \item \code{q.values} : blabla
}
}
\description{
This function allows the user to find groups of covarying loci on the RDA space.
}
\details{
Blablabla
}
\author{
Thibaut Capblancq
}
