% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loci_modules.R
\name{loci_modules}
\alias{loci_modules}
\title{Find groups of co-varying loci}
\usage{
loci_modules(RDA, nb_clusters)
}
\arguments{
\item{RDA}{a RDA model from which to extract loci and environmental variable scores}

\item{nb_clusters}{an integer specifying the number of cluster to identify}
}
\value{
A \code{list} containing :
\itemize{
  \item \code{loci} : a \code{data.frame} with the cluster associated with each locus
  \item \code{polygons} : polygon coordinates to outline the different cluster on a 2D RDA space 
}
}
\description{
The \code{loci_modules} function allows the user to find groups of covarying loci on the RDA space.
}
\details{
Identify a discrete number of adaptive groups across the 2D RDA space.
Next version will use more than two dimensions
}
\author{
Thibaut Capblancq
}
