% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CI_intsec_RD.R
\name{max_Q}
\alias{max_Q}
\title{Quantile of Maximum of Multivariate Normal Random Variable}
\usage{
max_Q(covmat, alpha, num_sim = 10^5)
}
\arguments{
\item{covmat}{covariance matrix of the multivariate normal random variable.}

\item{alpha}{desired upper quantile value.}

\item{num_sim}{number of simulations used to calculate the quantile;
the default is \code{10^5}.}
}
\value{
\eqn{(1 - \alpha)}th quantile of maximum of a multivariate normal random variable.
}
\description{
Calculates the quantile of maximum of a multivariate normal random variable,
by a Monte Carlo simulation method.
}
\examples{
covmat <- matrix(c(1, 0.5, 0.5, 1), 2, 2)
max_Q(covmat, 0.05)
}
