% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_calc_RD.R
\name{minb_fun}
\alias{minb_fun}
\title{Minimum Modulus}
\usage{
minb_fun(C_pair, X, mon_ind, swap = FALSE)
}
\arguments{
\item{C_pair}{a pair of smoothness parameters \eqn{(C, C')}.}

\item{X}{A data matrix.}

\item{mon_ind}{index number for monotone variables.}

\item{swap}{indicator for whether we take (C', C) instead of (C, C').}
}
\value{
the value of the smallest possible modulus value.
}
\description{
Calculates the smallest possible modulus value, i.e., \eqn{\omega(0)}.
}
\examples{
X <- matrix(c(1, -2, -3, 4, 5, -6), nrow = 3, ncol = 2)
mon_ind <- c(1, 2)
C_pair <- c(0.5, 1)
C_pair2 <- c(Inf, 1)
minb_fun(C_pair, X, mon_ind)
minb_fun(C_pair2, X, mon_ind)
}
