% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_calc.R
\name{mod_del_cal}
\alias{mod_del_cal}
\title{mod_del_cal function... see code for details (for now!)}
\usage{
mod_del_cal(gamvec, Cvec, Xt, Xc, mon_ind, sigma_t, sigma_c,
  alpha = 0.05)
}
\description{
mod_del_cal function... see code for details (for now!)
}
