% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_calc.R
\name{mod_del_cal_orc}
\alias{mod_del_cal_orc}
\title{mod_del_cal_orc}
\usage{
mod_del_cal_orc(gamma, C, maxgam, maxC, Xt, Xc, mon_ind, sigma_t, sigma_c,
  alpha = 0.05)
}
\arguments{
\item{gamma}{}

\item{C}{}

\item{maxgam}{}

\item{maxC}{}

\item{Xt}{}

\item{Xc}{}

\item{mon_ind}{}

\item{sigma_t}{}

\item{sigma_c}{}

\item{alpha}{}
}
\description{
mod_del_cal_orc
}
