% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_calc_RD.R
\name{modsol_RD}
\alias{modsol_RD}
\title{Modulus of Continuity for the RD Parameter}
\usage{
modsol_RD(delta, C_pair, Xt, Xc, mon_ind, sigma_t, sigma_c, swap = FALSE)
}
\arguments{
\item{delta}{a nonnegative value of \eqn{\delta}.}

\item{C_pair}{a pair of smoothness parameters \eqn{(C, C')}.}

\item{Xt}{\eqn{n_t} by \eqn{k} design matrix for the treated units.}

\item{Xc}{\eqn{n_c} by \eqn{k} design matrix for the control units.}

\item{mon_ind}{index number for monotone variables.}

\item{sigma_t}{standard deviation of the error term for the treated units
(either length 1 or \eqn{n_t}).}

\item{sigma_c}{standard deviation of the error term for the control units
(either length 1 or \eqn{n_c}).}

\item{swap}{indicator for whether we take (C', C) instead of (C, C').}
}
\value{
A list of four components. \code{bt} gives the modulus value
corresponding to the treated observations, while \code{delta_t} gives
the value of the inverse modulus corresponding to the treated observations.
\code{bc} and \code{delta_c} give the analogous values for the control observations.
}
\description{
Calculates the modulus of continuity for the RDD problem.
}
\examples{
n <- 500
d <- 2
X <- matrix(rnorm(n * d), nrow = n, ncol = d)
tind <- X[, 1] > 0 & X[, 2] > 0
Xt <- X[tind == 1, ,drop = FALSE]
Xc <- X[tind == 0, ,drop = FALSE]
C_pair <- c(0.1, 0.2)
mon_ind <- c(1, 2)
sigma <- rnorm(n)^2 + 1
sigma_t <- sigma[tind == 1]
sigma_c <- sigma[tind == 0]
modsol_RD(1, C_pair, Xt, Xc, mon_ind, sigma_t, sigma_c, swap = FALSE)
}
