% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_fcn_RD.R
\name{pmin_mat}
\alias{pmin_mat}
\title{Row-wise Minimum}
\usage{
pmin_mat(mat_var)
}
\arguments{
\item{mat_var}{input matrix}
}
\value{
a vecvor with a dimension \code{nrow(mat_var)}.
}
\description{
Calculates row-wise minimum of a matrix by looping over columns.
}
\details{
This function is useful when \code{nrow} is much greater than \code{ncol}.
}
