% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_fcn_RD.R
\name{pos}
\alias{pos}
\title{Positive Part}
\usage{
pos(X)
}
\arguments{
\item{X}{A data matrix.}
}
\value{
A matrix in the same dimension of \code{X}, after \code{max(x,0)}
was applied to each component \code{x} of \code{X}.
}
\description{
Calculates component-wise \code{max(X,0)} for a matrix X.
}
\examples{
X <- matrix(c(1, -2, -3, 4, 5, -6), nrow = 3, ncol = 2)
pos(X)
}
