% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdadapt.R
\name{rdadapt}
\alias{rdadapt}
\alias{rdadapt,rda,ANY,ANY-method}
\alias{rdadapt,missing,missing,data.frame-method}
\alias{rdadapt,missing,missing,matrix-method}
\title{RDA based genome scan}
\usage{
rdadapt(RDA, K, zscores)

\S4method{rdadapt}{rda,ANY,ANY}(RDA, K)

\S4method{rdadapt}{missing,missing,data.frame}(zscores)

\S4method{rdadapt}{missing,missing,matrix}(zscores)
}
\arguments{
\item{RDA}{blabla}

\item{K}{blabla}
}
\value{
A \code{data.frame} containing :
\itemize{
  \item \code{p.values} : blabla
  \item \code{q.values} : blabla
}
}
\description{
This function allows the user to ...
}
\details{
Blablabla
}
\author{
Thibaut Capblancq
}
