% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sim_fcns.R
\name{res_form}
\alias{res_form}
\title{Dataframe Generation After Simulations}
\usage{
res_form(res_all, trueC, method_name)
}
\arguments{
\item{res_all}{data matrix generated from simulations.}

\item{trueC}{a vector of true smoothness parameters used.}

\item{method_name}{a vector of names of CI construction methods used.}
}
\value{
a list containing coverage and length results stored as \code{data.frame}.
}
\description{
Dataframe Generation After Simulations
}
\examples{
trueC <- c(0.1, 0.2)
method_name <- c("a", "b")
m_len <- length(method_name)
trueC_len <- length(trueC)
nIters <- 50
rowlen <- m_len * trueC_len * 2
res_all <- matrix(stats::rnorm(rowlen * nIters), nrow = rowlen, ncol = nIters)
res_form(res_all, trueC, method_name)
}
