% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sim_fcns.R
\name{res_form_pow}
\alias{res_form_pow}
\title{Dataframe Generation After Simulations (Power Function)}
\usage{
res_form_pow(res_all, true_val_vec, method_name)
}
\arguments{
\item{res_all}{data matrix generated from simulations.}

\item{true_val_vec}{a vector of true values.}

\item{method_name}{a vector of names of CI construction methods used.}
}
\value{
a list containing power results.
}
\description{
Dataframe Generation After Simulations (Power Function)
}
\examples{
true_val_vec <- c(0.1, 0.2)
method_name <- c("a", "b")
m_len <- length(method_name)
true_val_len <- length(true_val_vec)
nIters <- 50
rowlen <- m_len * true_val_len
res_all <- matrix(stats::rnorm(rowlen * nIters), nrow = rowlen, ncol = nIters)
res_form(res_all, true_val_vec, method_name)
}
