% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sim_fcns.R
\name{res_gen}
\alias{res_gen}
\title{CI Result Generation}
\usage{
res_gen(
  obs_data,
  method_sym,
  C_len,
  mon_ind,
  alpha,
  trueC_min,
  trueC_max,
  true_val = 1
)
}
\arguments{
\item{obs_data}{a list object generated by \code{gen_obs} function.}

\item{method_sym}{a vector CI construction methods to be used; currently supports
\code{c("Ex", "rdr")}.}

\item{C_len}{the number of parameters to adapt to.}

\item{mon_ind}{the monotonicity index.}

\item{alpha}{the desired level of non-coverage probability.}

\item{trueC_min}{the value of the smoothness parameter for the smoothest true function}

\item{trueC_max}{the value of the smoothness parameter for the least smooth true function}

\item{true_val}{the true parameter value; default is 1.}
}
\value{
a vector of (coverage, CI_length) for each method - a vector of
length \code{len(method_sym) * 2}.
}
\description{
CI Result Generation
}
\examples{
obs_data <- gen_obs(100, 1, 1/2, "Lin", "het", "equi")
method_sym <- c("Ex", "rdr")
res_gen(obs_data, method_sym, 7, c(1), 0.05, 1/5, 1)
}
