% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_fcn_RD.R
\name{sd_w}
\alias{sd_w}
\title{Standard Deviation Given Weights}
\usage{
sd_w(w_t, w_c, sigma_t, sigma_c)
}
\arguments{
\item{w_t}{weights for treated observations.}

\item{w_c}{weights for control observations.}

\item{sigma_t}{conditional standard deviations for the treated.}

\item{sigma_c}{conditional standard deviations for the control.}
}
\value{
a scalar value.
}
\description{
Calculate a standard deviation of a linear estimator given weights and conditional
standard deviations.
}
\examples{
w_t <- rep(1/10, 10)
w_c <- rep(1/5, 5)
sigma_t <- rep(1/2, 10)
sigma_c <- rep(1/3, 5)
sd_w(w_t, w_c, sigma_t, sigma_c)
}
