% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sd_fcns.R
\name{sigmaNN2}
\alias{sigmaNN2}
\title{Nearest Neighborhood Estimation of Conditional Standard Deviation}
\usage{
sigmaNN2(Xt, Xc, Yt, Yc, N = 3)
}
\arguments{
\item{Xt}{\eqn{n_t} by \eqn{k} design matrix for the treated units.}

\item{Xc}{\eqn{n_c} by \eqn{k} design matrix for the control units.}

\item{Yt}{outcome value for the treated group observations.}

\item{Yc}{outcome value for the control group observations.}

\item{N}{the number of nearest neighbors; default is \code{N = 3}.}
}
\value{
a list containing conditional standard deviation estimates for treated
observations (\code{sigma.t}) and control observations (\code{sigma.c})
}
\description{
Standard deviation is calculated by the formula (23) of Armstrong and Kolesár (2020)
}
\details{
Currently, this is not being used, although this works for general dimensions;
debugging to be done.
}
\references{
{

\cite{Armstrong, Timothy B., and Michal Kolesár. 2020.
"Simple and honest confidence intervals in nonparametric regression."
Quantitative Economics 11 (1): 1–39.}

}
}
