% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sd_fcns.R
\name{sigmaSvm}
\alias{sigmaSvm}
\title{Standard Deviation Calculation by Silverman's Rule}
\usage{
sigmaSvm(Xt, Xc, Yt, Yc, t.dir = c("left", "right"))
}
\arguments{
\item{Xt}{\eqn{n_t} by \eqn{k} design matrix for the treated units.}

\item{Xc}{\eqn{n_c} by \eqn{k} design matrix for the control units.}

\item{Yt}{outcome value for the treated group observations.}

\item{Yc}{outcome value for the control group observations.}

\item{t.dir}{treatment direction; \code{t.dir = "left"} if \eqn{x < 0} is treated.
Otherwise, \code{t.dir = "right"}.}
}
\value{
a list containing conditional standard deviation estimates for treated
observations (\code{sigma.t}) and control observations (\code{sigma.c})
}
\description{
Standard deviation is calculated using Silverman's Rule,
as in RDHonest::NPRPrelimVar.fit
}
\details{
This works only for one-dimensional cases.
}
\examples{
n <- 500
d <- 1
X <- matrix(rnorm(n * d), nrow = n, ncol = d)
tind <- X[, 1] < 0
Xt <- X[tind == 1, ,drop = FALSE]
Xc <- X[tind == 0, ,drop = FALSE]
sigma <- rep(1, n)
sigma_t <- sigma[tind == 1]
sigma_c <- sigma[tind == 0]
Yt = 1 + rnorm(length(sigma_t), mean = 0, sd = sigma_t)
Yc = rnorm(length(sigma_c), mean = 0, sd = sigma_c)
sigmaSvm(Xt, Xc, Yt, Yc, "left")
}
