% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCres_comb.R
\name{spec.data.all}
\alias{spec.data.all}
\title{Data Generation for All Specifications}
\usage{
spec.data.all(spec.all, which.fun, n, th, C.big, C.small, sig.scale, seed = 1)
}
\arguments{
\item{spec.all}{\code{data.frame} object created by \code{spec.grid} function.}

\item{which.fun}{which regression function to evaluate; supports \code{1:6}.}

\item{n}{sample size.}

\item{th}{true RD parameter.}

\item{C.big}{the value of C corresponding to the large C.}

\item{C.small}{the value of C corresponding to the small C.}

\item{sig.scale}{scaling parameter for the standard deviation.}

\item{seed}{the random number seed used to generate regressors; the default is
\code{seed = 1}.}
}
\value{
a list with \code{nrow(spec.all)} components, where each of them contains
data corresponding to each specification.
}
\description{
Generates running variables, and corresponding standard deviations and
regression function values for all specifications.
}
\examples{
spec.all <- spec.grid(c("unif", "homo", "trueCsmall"))
spec.data.all(spec.all, 4, 100, 1, 3, 1, 1/2)
spec.data.all(spec.all, 5, 100, 1, 3, 1, 1/2)
}
