% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCres_comb.R
\name{spec.data}
\alias{spec.data}
\title{Data Generation for a Given Specification}
\usage{
spec.data(spec.ind, which.fun, n, th, C.big, C.small, sig.scale, seed = 1)
}
\arguments{
\item{spec.ind}{\code{data.frame} with a single rows, with \code{names(spec.ind)} containing
\code{c("unif", "homo", "trueCsmall")}.}

\item{which.fun}{which regression function to evaluate; supports \code{1:6}.}

\item{n}{sample size.}

\item{th}{true RD parameter.}

\item{C.big}{the value of C corresponding to the large C.}

\item{C.small}{the value of C corresponding to the small C.}

\item{sig.scale}{scaling parameter for the standard deviation.}

\item{seed}{the random number seed used to generate regressors; the default is
\code{seed = 1}.}
}
\value{
a list with components \code{x} (running variables),
\code{sig} (standard deviations), and \code{fx} (regression function values).
}
\description{
Generates running variables, and corresponding standard deviations and
regression function values given a specification.
}
\examples{
spec.ind <- data.frame(unif = 1, homo = 0, trueCsmall = 0)
spec.data(spec.ind, 1, 100, 1, 3, 1, 1/2)
spec.data(spec.ind, 6, 100, 1, 3, 1, 1/2)
}
