% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCres_comb.R
\name{spec.grid}
\alias{spec.grid}
\title{Specification Data Frame Generation}
\usage{
spec.grid(spec)
}
\arguments{
\item{spec}{a vector of strings, containing specification information.}
}
\value{
\code{data.frame} object containing binary indicators for all possible
combinations of specifications.
}
\description{
Generates all possible combinations of specifications.
}
\examples{
spec.grid(c("unif", "homo"))
}
