% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sim_fcns.R
\name{spec_set}
\alias{spec_set}
\title{Specification Setting for Simulation}
\usage{
spec_set(spec_num, trueC_len, trueC_sc = NULL)
}
\arguments{
\item{spec_num}{a number symbolizing the specification to be used.}

\item{trueC_len}{the number of true smoothness spaces.}

\item{trueC_sc}{a scalar value of the smoothness parameter,
used for power function calculations.}
}
\value{
a list of elements.
}
\description{
Specification Setting for Simulation
}
\examples{
spec_set(1, 5)
spec_set(2, 5)
spec_set(1, 1, 0.5)
}
