% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CI_const_RD.R
\name{sup_bias_Lhat_RD}
\alias{sup_bias_Lhat_RD}
\title{Worst-case Bias of Estimator}
\usage{
sup_bias_Lhat_RD(delta, Cj, Cbar, Xt, Xc, mon_ind, sigma_t, sigma_c, ht, hc)
}
\arguments{
\item{delta}{a nonegative scalar value:
it can be left unspecified if \code{ht} and \code{hc} are specified.}

\item{Cj}{the smoothness parameter aiming to adapt to.}

\item{Cbar}{the largest smoothness parameter.}

\item{Xt}{\eqn{n_t} by \eqn{k} design matrix for the treated units.}

\item{Xc}{\eqn{n_c} by \eqn{k} design matrix for the control units.}

\item{mon_ind}{index number for monotone variables.}

\item{sigma_t}{standard deviation of the error term for the treated units
(either length 1 or \eqn{n_t}).}

\item{sigma_c}{standard deviation of the error term for the control units
(either length 1 or \eqn{n_c}).}

\item{ht}{the modulus value for the treated observations;
it can be left unspecified if \code{delta} is specified.}

\item{hc}{the modulus value for the control observations;
it can be left unspecified if \code{delta} is specified.}
}
\value{
a scalar value
}
\description{
Calculates the worst case bias of the estimator
\eqn{Lhat_j(\delta)}.
}
\examples{
n <- 500
d <- 2
X <- matrix(rnorm(n * d), nrow = n, ncol = d)
tind <- X[, 1] > 0 & X[, 2] > 0
Xt <- X[tind == 1, ,drop = FALSE]
Xc <- X[tind == 0, ,drop = FALSE]
mon_ind <- c(1, 2)
sigma <- rnorm(n)^2 + 1
sigma_t <- sigma[tind == 1]
sigma_c <- sigma[tind == 0]
sup_bias_Lhat_RD(1, 1/2, 1, Xt, Xc, mon_ind, sigma_t, sigma_c)
}
