% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CI_intsec_RD.R
\name{tau_sol}
\alias{tau_sol}
\title{Coverage Probability Calculation for Intersection CI}
\usage{
tau_sol(
  Cvec,
  Xt,
  Xc,
  mon_ind,
  sigma_t,
  sigma_c,
  alpha,
  num_sim = 10^4,
  delta_init = 1.96
)
}
\arguments{
\item{Cvec}{a sequence of smoothness parameters}

\item{Xt}{\eqn{n_t} by \eqn{k} design matrix for the treated units.}

\item{Xc}{\eqn{n_c} by \eqn{k} design matrix for the control units.}

\item{mon_ind}{index number for monotone variables.}

\item{sigma_t}{standard deviation of the error term for the treated units
(either length 1 or \eqn{n_t}).}

\item{sigma_c}{standard deviation of the error term for the control units
(either length 1 or \eqn{n_c}).}

\item{alpha}{desired upper quantile value.}

\item{num_sim}{number of simulations used to calculate the quantile;
the default is \code{10^4}.}

\item{delta_init}{the value of \eqn{\delta} to be used in simulating the quantile;
theoretically, its value does not matter asymptotically.}
}
\value{
a list of two values, \code{del_sol}, the proper value of \eqn{\delta} to be used
for individual CIs, and \code{tau_sol}, the non-coverage probability associated with the
value of \eqn{\delta}.
}
\description{
Calculates the coverage probability of individual confidence intervals
to ensure the proper coverage of the intersection confidence interval.
}
\examples{
n <- 500
d <- 2
X <- matrix(rnorm(n * d), nrow = n, ncol = d)
tind <- X[, 1] > 0 & X[, 2] > 0
Xt <- X[tind == 1, ,drop = FALSE]
Xc <- X[tind == 0, ,drop = FALSE]
mon_ind <- c(1, 2)
sigma <- rnorm(n)^2 + 1
sigma_t <- sigma[tind == 1]
sigma_c <- sigma[tind == 0]
tau_sol((1:5)/5, Xt, Xc, mon_ind, sigma_t, sigma_c, 0.05)
}
