% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_fcn_RD.R
\name{Vplus}
\alias{Vplus}
\title{Function for \eqn{(x)_{V+}}}
\usage{
Vplus(X, mon_ind)
}
\arguments{
\item{X}{A data matrix.}

\item{mon_ind}{index number for monotone variables.}
}
\value{
A matrix in the same dimension of \code{X},
after \eqn{(x)_{V+}} was applied to each row \eqn{x} of \code{X}.
}
\description{
Calculates rowwise \eqn{(x)_{V+}} for each row \eqn{x}
of a matrix \code{X}.
}
\examples{
X <- matrix(c(1, -2, -3, 4, 5, -6), nrow = 3, ncol = 2)
mon_ind <- c(1, 2)
Vplus(X, mon_ind)
}
