% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CI_const.R
\name{which_proc}
\alias{which_proc}
\title{(One-sided) Procedure choice}
\usage{
which_proc(Cgrid, gamma, C, Xt, Xc, mon_ind, sigma_t, sigma_c, lower,
  alpha, procmat, ProcNames, Nsim_proc, rhofun = c("diff", "ratio"))
}
\arguments{
\item{gamma}{length two vector of gammas (\eqn{(\gamma_1, \gamma_2)'})}

\item{C}{length two vector of Cs (\eqn{(C_1, C_2)'})}

\item{Xt}{\eqn{n_t \times k} design matrix for the treated units}

\item{Xc}{\eqn{n_c \times k} design matrix for the control units}

\item{mon_ind}{indice of the monotone variables}

\item{sigma_t}{standard deviation of the error term for the treated units
(either length 1 or \eqn{n_t})}

\item{sigma_c}{standard deviation of the error term for the control units}

\item{lower}{}

\item{alpha}{}

\item{rhofun}{corresponds to \eqn{\rho} in our documentation}
}
\description{
Calculates the adaptive procedure "closest" to the oracle one
}
