% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adaptive_groups.R
\name{adaptive_groups}
\alias{adaptive_groups}
\alias{adaptive_groups,rda,ANY,missing-method}
\alias{adaptive_groups,rda,ANY,SpatRaster-method}
\title{Site clustering}
\usage{
adaptive_groups(RDA, K, env, nb_clusters)

\S4method{adaptive_groups}{rda,ANY,missing}(RDA, K, nb_clusters)

\S4method{adaptive_groups}{rda,ANY,SpatRaster}(RDA, K, env, nb_clusters)
}
\arguments{
\item{RDA}{blabla}

\item{K}{blabla}
}
\value{
A \code{data.frame} containing :
\itemize{
  \item \code{p.values} : blabla
  \item \code{q.values} : blabla
}
}
\description{
This function allows the user to find groups of covarying 
sites/individuals/populations within the RDA space.
}
\details{
Blablabla
}
\author{
Thibaut Capblancq
}
