% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adaptive_index.R
\name{adaptive_index}
\alias{adaptive_index}
\alias{adaptive_index,rda,ANY,SpatRaster,ANY-method}
\alias{adaptive_index,rda,ANY,data.frame,missing-method}
\title{Adaptive index}
\usage{
adaptive_index(RDA, K, env, env_mask = NULL, method = "loadings")

\S4method{adaptive_index}{rda,ANY,SpatRaster,ANY}(RDA, K, env, env_mask = NULL, method = "loadings")

\S4method{adaptive_index}{rda,ANY,data.frame,missing}(RDA, K, env, method = "loadings")
}
\arguments{
\item{RDA}{blabla}

\item{K}{blabla}

\item{env}{blabla}

\item{env_mask}{(\emph{optional, default} \code{NULL}) \cr blabla}

\item{method}{(\emph{default} \code{'loadings'}) \cr blabla}

\item{scale_env}{blabla}

\item{center_env}{blabla}
}
\value{
A \code{matrix} containing blabla
}
\description{
This function allows the user to project the adaptive component turnover across 
a landscape.
}
\details{
Blablabla
}
\author{
Thibaut Capblancq
}
