% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CI_mm.R
\name{CI_length_RD}
\alias{CI_length_RD}
\title{Minimum Half-length Function}
\usage{
CI_length_RD(b, C, Xt, Xc, mon_ind, sigma_t, sigma_c, alpha)
}
\arguments{
\item{b}{point where the inverse modulus is evaluated at.}

\item{C}{a scalar smoothness parameter.}

\item{Xt}{\eqn{n_t} by \eqn{d} design matrix for the treated units.}

\item{Xc}{\eqn{n_c} by \eqn{d} design matrix for the control units.}

\item{mon_ind}{index number for monotone variables.}

\item{sigma_t}{standard deviation of the error term for the treated units
(either length 1 or \eqn{n_t}).}

\item{sigma_c}{standard deviation of the error term for the control units
(either length 1 or \eqn{n_c}).}

\item{alpha}{the desired level of non-coverage probability.}
}
\value{
the minimum half-length corresponding to the modulus value \code{b}
}
\description{
Calculates the minimum half-length, as a function of the modulus value.
}
\details{
In the paper, the minimum half-length is expressed as a function of
\eqn{\delta};
here, we use the relationship that \eqn{\delta} is the inverse modulus value
given \eqn{b}.
The returned value will be minimized with respect to \eqn{b} later
}
\examples{
X <- matrix(rnorm(500 * 2), nrow = 500, ncol = 2)
tind <- X[, 1] > 0 & X[, 2] > 0
Xt <- X[tind == 1, ,drop = FALSE]
Xc <- X[tind == 0, ,drop = FALSE]
sigma <- rnorm(500)^2 + 1
sigma_t <- sigma[tind == 1]
sigma_c <- sigma[tind == 0]
CI_length_RD(1.1, 1/2, Xt, Xc, c(1, 2), sigma_t , sigma_c, 0.05)
}
