% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CI_const.R
\name{CI_length_RD}
\alias{CI_length_RD}
\title{Length of minimax CI for the RDD}
\usage{
CI_length_RD(b, gamma, C, Xt, Xc, mon_ind, sigma_t = 1, sigma_c = 1,
  alpha = 0.05)
}
\arguments{
\item{b}{point where the inverse modulus is evaluated at}

\item{gamma}{length two vector of gammas (\eqn{(\gamma_1, \gamma_2)'})}

\item{C}{length two vector of Cs (\eqn{(C_1, C_2)'})}

\item{Xt}{\eqn{n_t \times k} design matrix for the treated units}

\item{Xc}{\eqn{n_c \times k} design matrix for the control units}

\item{mon_ind}{indice of the monotone variables}

\item{sigma_t}{standard deviation of the error term for the treated units
(either length 1 or \eqn{n_t})}

\item{sigma_c}{standard deviation of the error term for the control units}

\item{alpha}{significance level}
}
\description{
Calculates the length of the minimax CI for the RDD
}
