% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CI_const.R
\name{CI_minimax_RD}
\alias{CI_minimax_RD}
\title{Minimax CI for the RD parameter}
\usage{
CI_minimax_RD(Yt, Yc, Xt, Xc, gam_min, C_max, mon_ind, sigma_t, sigma_c,
  alpha = 0.05, modres)
}
\arguments{
\item{Yt}{}

\item{Yc}{}

\item{Xt}{}

\item{Xc}{}

\item{gam_min}{}

\item{C_max}{}

\item{mon_ind}{}

\item{sigma_t}{}

\item{sigma_c}{}

\item{alpha}{}

\item{modres}{}
}
\description{
This function calculates shortest minimax (fixed-length) CI for the RD
parameter
}
