% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CI_const.R
\name{CI_one_sd_RD}
\alias{CI_one_sd_RD}
\title{(One-sided) Adaptive CI for RDD}
\usage{
CI_one_sd_RD(bpairmat, dmat, gamma, C, Xt, Xc, mon_ind, sigma_t, sigma_c,
  Yt, Yc, lower, alpha = 0.05)
}
\arguments{
\item{bpairmat}{}

\item{dmat}{}

\item{gamma}{}

\item{C}{}

\item{Xt}{}

\item{Xc}{}

\item{mon_ind}{}

\item{sigma_t}{}

\item{sigma_c}{}

\item{Yt}{}

\item{Yc}{}

\item{lower}{}

\item{alpha}{}
}
\description{
Calculates the (one-sided) adaptive CI for RDD
}
