% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CI_mm.R
\name{cv_a}
\alias{cv_a}
\title{\eqn{cv_\alpha(t)} Function}
\usage{
cv_a(t, alpha)
}
\arguments{
\item{t}{the non-centrality parameter.}

\item{alpha}{probability}
}
\value{
the quantile value
}
\description{
\eqn{cv_\alpha(t)} Function
}
\examples{
cv_a(1, 0.05)
}
