% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sim_fcns.R
\name{gen_obs}
\alias{gen_obs}
\title{Sample Generation for Simulations}
\usage{
gen_obs(
  n,
  d,
  C,
  spec = c("Lin"),
  sd_spec = c("hom", "het"),
  X_dist = c("equi"),
  true_val = 1
)
}
\arguments{
\item{n}{sample size.}

\item{d}{number of running variables.}

\item{C}{a smoothness parameter.}

\item{spec}{specifications to use.}

\item{sd_spec}{homoskedastic error if \code{sd_spec = "hom"};
heteroskedastic if \code{sd_spec = "het"}.}

\item{X_dist}{distribution of running variables}

\item{true_val}{the true parameter value; the default is 1.}
}
\value{
a list with \code{Yt}, \code{Yc}, \code{Xt}, \code{Xc}, \code{sigma_t},
and \code{sigma_c}.
}
\description{
Sample Generation for Simulations
}
\examples{
gen_obs(100, 1, 1/2, "Lin", "het", "equi")
}
