% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomic_offset.R
\name{genomic_offset}
\alias{genomic_offset}
\alias{genomic_offset,rda,ANY,SpatRaster,ANY,ANY,missing-method}
\alias{genomic_offset,rda,ANY,data.frame,ANY,missing,missing-method}
\alias{genomic_offset,rda,ANY,data.frame,missing,missing,ANY-method}
\title{Genomic offset}
\usage{
genomic_offset(
  RDA,
  K,
  env_pres,
  env_fut,
  env_mask = NULL,
  env_gar,
  method = "loadings"
)

\S4method{genomic_offset}{rda,ANY,SpatRaster,ANY,ANY,missing}(RDA, K, env_pres, env_fut, env_mask = NULL, method = "loadings")

\S4method{genomic_offset}{rda,ANY,data.frame,ANY,missing,missing}(RDA, K, env_pres, env_fut, method = "loadings")

\S4method{genomic_offset}{rda,ANY,data.frame,missing,missing,ANY}(RDA, K, env_pres, env_gar, method = "loadings")
}
\arguments{
\item{RDA}{blabla}

\item{K}{blabla}

\item{env_pres}{blabla}

\item{env_fut}{blabla}

\item{env_mask}{(\emph{optional, default} \code{NULL}) \cr blabla}

\item{method}{(\emph{default} \code{'loadings'}) \cr blabla}
}
\value{
A \code{list} containing blabla
}
\description{
This function allows the user to predict genomic offset from a RDA model.
}
\details{
Blablabla
}
\author{
Thibaut Capblancq
}
