% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_calc.R
\name{invmod}
\alias{invmod}
\title{Invsere modulus}
\usage{
invmod(b, gamma, C, X, mon_ind, sigma = 1)
}
\arguments{
\item{b}{point where the inverse modulus is evaluated at}

\item{gamma}{length two vector of gammas}
}
\description{
Calculates the inverse modulus for the regression function at a point
problem. More specifically, this calcultes
\eqn{\omega^{-1}(b, \Lambda(\gamma_2, C_2),\Lambda(\gamma_1, C_1)) }
}
