% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_calc.R
\name{invmod_RD}
\alias{invmod_RD}
\title{Invsere modulus for RDD}
\usage{
invmod_RD(b, gamma, C, Xt, Xc, mon_ind, sigma_t = 1, sigma_c = 1,
  ret_list = FALSE)
}
\arguments{
\item{b}{point where the inverse modulus is evaluated at}

\item{gamma}{length two vector of gammas (\eqn{(\gamma_1, \gamma_2)'})}

\item{C}{length two vector of Cs (\eqn{(C_1, C_2)'})}

\item{Xt}{\eqn{n_t \times k} design matrix for the treated units}

\item{Xc}{\eqn{n_c \times k} design matrix for the control units}

\item{mon_ind}{indice of the monotone variables}

\item{sigma_t}{standard deviation of the error term for the treated units
(either length 1 or \eqn{n_t})}

\item{sigma_c}{standard deviation of the error term for the control units}

\item{ret_list}{indicator of whether to return a more detailed list of results}
}
\description{
Calculates the inverse modulus for the RDD
problem.
}
