% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_fcn_RD.R
\name{K_fun}
\alias{K_fun}
\title{Kernel Function for Adaptive Procedures}
\usage{
K_fun(b, C_pair, X, mon_ind, swap = FALSE)
}
\arguments{
\item{b}{"bandwidth" parameter.}

\item{C_pair}{a pair of smoothness parameters \eqn{(C, C')}}

\item{X}{data matrix.}

\item{mon_ind}{index number for monotone variables.}

\item{swap}{we take (C', C) instead of (C, C') if \code{swap = TRUE}.}
}
\value{
A numeric vector with the length equal to nrow(X);
if b = 0, returns 0 vector.
}
\description{
Calculates \eqn{K(x/b; C, C')}.
}
\examples{
X <- matrix(c(1, -2, -3, 4, 5, -6), nrow = 3, ncol = 2)
mon_ind <- c(1, 2)
C_pair <- c(0.5, 1)
K_fun(1, C_pair, X, mon_ind)
}
