% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cvec_choice.R
\name{l_adpt}
\alias{l_adpt}
\title{Worst-case Excess Length of the Adaptive Procedure}
\usage{
l_adpt(
  Cpr,
  Cvec,
  C,
  Xt,
  Xc,
  mon_ind,
  sigma_t,
  sigma_c,
  alpha,
  sim = 10^5,
  tau_res,
  hmat
)
}
\arguments{
\item{Cpr}{\eqn{C'} in the paper}

\item{Cvec}{a sequence of smoothness parameters}

\item{C}{the Lipschitz coefficient for the function space we consider}

\item{Xt}{\eqn{n_t} by \eqn{k} design matrix for the treated units.}

\item{Xc}{\eqn{n_c} by \eqn{k} design matrix for the control units.}

\item{mon_ind}{index number for monotone variables.}

\item{sigma_t}{standard deviation of the error term for the treated units
(either length 1 or \eqn{n_t}).}

\item{sigma_c}{standard deviation of the error term for the control units
(either length 1 or \eqn{n_c}).}

\item{alpha}{desired upper quantile value.}

\item{sim}{number of simulated observations to calculate the
expectation of the minimum of multivariate normal random variables.}

\item{tau_res}{a list produced by \code{tau_calc}; can be left unspecified}

\item{hmat}{a matrix of bandwidths to be used in the adaptive procedure;
can be left unspecified}
}
\description{
Worst-case Excess Length of the Adaptive Procedure
}
\examples{
n <- 500
d <- 2
X <- matrix(rnorm(n * d), nrow = n, ncol = d)
tind <- X[, 1] < 0 & X[, 2] < 0
Xt <- X[tind == 1, ,drop = FALSE]
Xc <- X[tind == 0, ,drop = FALSE]
mon_ind <- c(1, 2)
sigma <- rnorm(n)^2 + 1
sigma_t <- sigma[tind == 1]
sigma_c <- sigma[tind == 0]
l_adpt(1/4, (1:5)/5, 2, Xt, Xc, mon_ind, sigma_t, sigma_c, 0.05)
}
