% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdadapt_data.R
\docType{data}
\name{maladapt}
\alias{maladapt}
\title{Maladaptation tutorial}
\format{
A \code{list} object with 3 elements:
\describe{
  \item{pop}{a \code{vector} identifying population for 2036 individuals}
  \item{coord}{a \code{data.frame} containing coordinates for 2036 individuals}
  \item{Y}{a \code{data.frame} containing genomic information for 2036 
  individuals about 2333 genes}
  \item{xx.current}{a \code{data.frame} containing 4 current environmental 
variables for 2036 individuals}
  \item{xx.future}{a \code{data.frame} containing 4 future environmental 
variables for 2036 individuals}
  \item{causal1}{a \code{vector} identifying ...}
  \item{causal2}{a \code{vector} identifying ...}
  \item{fitness}{a \code{data.frame} containing ...}
}
}
\usage{
maladapt
}
\description{
A \code{list} containing names, coordinates and genomic informations for 
\pkg{rdadapt} tutorial.
}
\keyword{datasets}
