% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_calc.R
\name{minb_fun}
\alias{minb_fun}
\title{Calculate omega(0)}
\usage{
minb_fun(gamma, C, X, mon_ind, swap = FALSE)
}
\arguments{
\item{gamma}{}

\item{C}{}

\item{X}{}

\item{mon_ind}{}

\item{swap}{}
}
\description{
Calculates the smallest possible b value, or omega(0)
}
