% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_calc.R
\name{mod_del_cal}
\alias{mod_del_cal}
\title{Modulus for adaptive RDD}
\usage{
mod_del_cal(gamvec, Cvec, Xt, Xc, mon_ind, sigma_t, sigma_c,
  alpha = 0.05)
}
\arguments{
\item{gamvec}{}

\item{Cvec}{}

\item{Xt}{}

\item{Xc}{}

\item{mon_ind}{}

\item{sigma_t}{}

\item{sigma_c}{}

\item{alpha}{}
}
\description{
Modulus for adaptive RDD
}
