% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_calc.R
\name{modsol_RD}
\alias{modsol_RD}
\title{Calculate modulus for RDD}
\usage{
modsol_RD(delta, gamma, C, Xt, Xc, mon_ind, sigma_t, sigma_c,
  swap = FALSE)
}
\arguments{
\item{delta}{}

\item{gamma}{}

\item{C}{}

\item{Xt}{}

\item{Xc}{}

\item{mon_ind}{}

\item{sigma_t}{}

\item{sigma_c}{}
}
\description{
Calculate modulus for RDD
}
