% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cvec_choice.R
\name{pmin_mat}
\alias{pmin_mat}
\title{Row-wise minimum}
\usage{
pmin_mat(mat_var)
}
\arguments{
\item{mat_var}{input matrix}
}
\value{
a vecvor with a dimension \code{nrow(mat_var)}.
}
\description{
Row-wise minimum
}
