% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdadapt.R
\name{rdadapt}
\alias{rdadapt}
\alias{rdadapt,rda,ANY,ANY-method}
\alias{rdadapt,missing,missing,data.frame-method}
\alias{rdadapt,missing,missing,matrix-method}
\title{Detect outlier loci based on their contribution to a RDA model}
\usage{
rdadapt(RDA, K, scores)

\S4method{rdadapt}{rda,ANY,ANY}(RDA, K)

\S4method{rdadapt}{missing,missing,data.frame}(scores)

\S4method{rdadapt}{missing,missing,matrix}(scores)
}
\arguments{
\item{RDA}{a RDA model from which to extract loci loadings}

\item{K}{an integer specifying the number of RDA axes to use for the detection}
}
\value{
A \code{data.frame} containing :
\itemize{
  \item \code{p.values} : the p-value associated with each locus
  \item \code{q.values} : the q-value associated with each locus, estimated using the \code{q.values} package and allowing to use a FDR (False Discovery Rate) approach instead of a p-value threshold to identify outliers.
}
}
\description{
The \code{rdadapt} function performs redundancy analysis and computes p-values to
test for outliers based on loci extremeness along a distribution of Mahalanobis distances 
estimated between each locus and the center of the RDA space using a certain number of axes (K). 
\code{rdadapt} accommodates individual genotypes or allele frequencies.
}
\details{
First, Mahalanobis distances are computed for all genetic marker using a robust estimate of both mean and covariance matrix between the \code{K} RDA vectors of loadings.
Then, to compute p-values, Mahalanobis distances are divided by a genomic inflation factor (\code{gif}), giving a scaled statistic that should follow a chi-squared distribution with \code{K} degrees of freedom.
}
\author{
Thibaut Capblancq
}
