% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sim_fcns.R
\name{res_gen}
\alias{res_gen}
\title{CI Result Generation}
\usage{
res_gen(n, spec_list, method_sym, C_len, true_val = 1)
}
\arguments{
\item{n}{sample size.}

\item{spec_list}{a list generated by \code{spec_set}.}

\item{method_sym}{a vector CI construction methods to be used; currently supports
\code{c("Ex", "rdr")}.}

\item{C_len}{the number of parameters to adapt to.}

\item{true_val}{the true parameter value; default is 1.}
}
\value{
a vector of (coverage, CI_length) for each method - a vector of
length \code{len(method_sym) * 2}.
}
\description{
CI Result Generation
}
\examples{
n <- 100
spec_list <- rdadapt::spec_set(1, 2)
method_sym <- c("Ex", "rdr")
res_gen(n, spec_list, method_sym, 5)
}
