% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sim_fcns.R
\name{res_gen}
\alias{res_gen}
\title{CI Result Generation}
\usage{
res_gen(
  n,
  spec_list,
  method_sym,
  C_len,
  Csvtv_const,
  lower = FALSE,
  true_val = 1,
  power = FALSE
)
}
\arguments{
\item{n}{sample size.}

\item{spec_list}{a list generated by \code{spec_set}.}

\item{method_sym}{a vector CI construction methods to be used; currently supports
\code{c("Ex", "rdr")}.}

\item{C_len}{the number of smoothness parameters to adapt to.}

\item{Csvtv_const}{the constant which governs the degree of conservativeness of the
adaptive and minimax procedures.}

\item{lower}{\code{TRUE} in order to generate a lower one-sided confidence interval}

\item{true_val}{the true parameter value; default is 1. It can be a vector for
power function calculations.}

\item{power}{generates power function result if \code{TRUE}.}
}
\value{
a vector of (coverage, CI_length) for each method - a vector of
length \code{len(method_sym) * 2}.
}
\description{
CI Result Generation
}
\examples{
n <- 100
spec_list <- rdadapt::spec_set(1, 2, 0.5)
method_sym <- c("Ex", "rdr")
res_gen(n, spec_list, method_sym, 5, 2)
res_gen(n, spec_list, method_sym, 5, 2, FALSE, c(0, 1), TRUE)
}
